requirements.txt
amqp==5.0.5
boto==2.43.0
boto3==1.17.33
botocore==1.20.33
celery==5.0.5
Flask==1
Flask-Babel==0.9
Flask-Compress==1.3.0
Flask-Cors==2.1.0
Flask-Login==0.3.0
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-User==0.6.8
Flask-WTF==0.12
flower==0.8.3
futures==3.0.3
psycopg2-binary==2.8.4
#psycopg2==2.8.6
pandas
pyyaml
haikunator
xlrd
chardet
csvkit==1.0.1
scipy
statsmodels
scikit-learn
gunicorn
